# public/apiDocsV1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "book-directory",
    "description": "",
    "contact": {
      "email": "adityatest821@gmail.com"
    },
    "version": "1.0.0",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/book/list/allAvailableBooks": {
      "get": {
        "deprecated": false,
        "tags": [
          "Book"
        ],
        "description": "List of all the Books",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "formData",
            "description": "page",
            "required": false,
            "type": "number"
          },
          {
            "name": "limit",
            "in": "formData",
            "description": "limit",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    "/book/addBookToStore": {
      "post": {
        "deprecated": false,
        "tags": [
          "Book"
        ],
        "description": "Add a Book",
        "consumes": [
          "application/x-www-form-urlencoded",
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "bookName",
            "in": "formData",
            "description": "bookName",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorName",
            "in": "formData",
            "description": "authorName",
            "required": true,
            "type": "string"
          },
          {
            "name": "isbnNum",
            "in": "formData",
            "description": "isbnNum",
            "required": true,
            "type": "string"
          },
          {
            "name": "costPrice",
            "in": "formData",
            "description": "costPrice",
            "required": true,
            "type": "string"
          },
          {
            "name": "status",
            "in": "formData",
            "description": "// 0 = inactive, 1 = active",
            "required": true,
            "type": "number"
          },
          {
            "name": "isDeleted",
            "in": "formData",
            "description": "// 0 = not-deleted, 1 = deleted",
            "required": true,
            "type": "number"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "description",
            "required": true,
            "type": "string"
          },
          {
            "name": "bookImage",
            "in": "formData",
            "description": "bookImage",
            "required": false,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    "/book/deleteBookFromStore": {
      "delete": {
        "deprecated": false,
        "tags": [
          "Book"
        ],
        "description": "Delete a Book from store",
        "consumes": [
          "application/x-www-form-urlencoded",
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "bookId",
            "in": "formData",
            "description": "bookId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
...
    "/book/list/allAuthors": {
      "get": {
        "deprecated": false,
        "tags": [
          "Book"
        ],
        "description": "List of all the Books",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "number"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "limit",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    "/book/list/authorWiseBook": {
      "get": {
        "deprecated": false,
        "tags": [
          "Book"
        ],
        "description": "List of all the Books",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorName",
            "in": "query",
            "description": "authorName",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "formData",
            "description": "page",
            "required": false,
            "type": "number"
          },
          {
            "name": "limit",
            "in": "formData",
            "description": "limit",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    "/book/bulkAddBooksToStore": {
      "post": {
        "deprecated": false,
        "tags": [
          "Book"
        ],
        "description": "Add a Book",
        "consumes": [
          "application/x-www-form-urlencoded",
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "uploadFile",
            "in": "formData",
            "description": "uploadFile",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    }
  }
}